Claude helped me point out some bugs.

+++ Simplified/Classes/Camera.py
@@ -162,8 +162,8 @@ class Camera:
 
         img_resized, _, left, top = self._letterbox(img_rgb, (640, 640))
 
-        img_input = np.expand_dims(img_resized, axis=0)  # shape (1, H, W, 3)
-        img_input = np.ascontiguousarray(img_input, dtype=np.float32)
+        img_input = np.expand_dims(img_resized, axis=0) # (1, 640, 640, 3)
+        img_input = np.ascontiguousarray(img_input, dtype=np.float16)
         return img_input
         
     def release(self):
